fix: I hate Python and Torch and NumPy so much.
This ecosystem insists on overloading their functions so that they do
something for every tensor dimension, rather than just forcing the user
to explicitly declare their expected dimensions and behavior. It's a
completely unnecessary footgun, trading a huge amount of safety and
correctness for saving a few keystrokes.

The problem was in the gradient calculation. In the continuous inverted
pendulum problem, the action that the environment expects is a
one-element array. You cannot simply pass it the action directly, it
must be in a one-element array. I assume this is some annoying quirk of
MuJoCo.

So I made the neural net spit out a one-element array instead of a
scalar. Fine. However, this means that when recording the actions of an
episode, I was accidentally recording a list of _one-element arrays_,
not a list of scalars. Let's say the epoch contained N steps. Then our
list of actions taken becomes an (N,1) tensor instead of an (N,)
tensor.

For some reason, torch.Distribution().log_prob accepts (N,1) tensors as
an argument! Why??? When would I ever want to do this??? What does that
even _mean_? But it does, and it also interprets their semantic meaning
to be something completely different from (N,) tensors. With (N,)
observations and (N,) actions, the distribution gives me back an (N,) of
the log-likelihood of each action given each observation (element-wise).
This is sane and reasonable. But then when I give it (N,1) actions with
the same observations, for some reason it spits out an (N,N) tensor! I
think this is log_prob for each of the (N,) observations evaluated at
each of the (N,) values, constructing a matrix of pairwise (rather than
element-wise) evaluations. Why??? Why would you design an API this way??
When would it ever make sense for the function be overloaded with
dramatic semantics changes rather than just throwing an error and
pointing the user at a different function???? But of course this is
still the correct shape to do a multiplication against the (N,)
weights, except here too it's now a matrix multiplication instead of an
element-wise multiplication!

Obviously, the resulting expression is some garbage that's not the
gradient, and optimizing it does not produce any useful behavior. In
fact, this buggy gradient calculation seems to cause the agent to
anti-learn. Instead of just putzing around randomly, it seems to
actively learn to take actions that produce less return. I have not
investigated why this is, and I don't care to.

I only discovered this by running both the broken implementation and
previous implementation side-by-side, running both for 2 epochs with
very small batch sizes (30), and seeing exactly which inputs to the
update differed (printing out the observations tensor, the actions
tensor, the weights tensor, and the result of the log_prob calculation).

Along the way, some other changes:

- Fixed an issue where the seed wasn't being set before model
  initialization, so initial model weights would be inconsistent between
  runs with allegedly the same seed. This is what finally enabled the
  side-by-side debugging.
- Narrow a bunch of types to make things simpler. Fuck it, everything is
  a float32 now.
- Refactor "gradient functions" into "weight functions", since the
  gradient calculation is the same for all of them.
- Refactor from collecting data episode-wise to column-wise (i.e. a list
  of observations, a list of actions, and a list of rewards, as opposed
  to a list of (observation, action, reward)). The increased correctness
  just wasn't worth the performance cost.